1. Static context refactoring 2. Better fix for hashjoins problem 3. Various other bug fixes 4. Eliminating some xqpString uses.

+++ src/util/dynamic_bitset.cpp
@@ -24,7 +24,7 @@ namespace zorba {
 std::ostream& operator <<(std::ostream& s, const DynamicBitset& set)
 {
   s << "BitSet[" << set.size() << "] = {";
-  for(int i = 0; i < set.size(); ++i) 
+  for(ulong i = 0; i < set.size(); ++i) 
   {
     if (set.get(i)) {
       s << i << ", ";
