+++ src/util/dynamic_bitset.h
@@ -34,13 +34,13 @@ private:
   typedef std::vector<uint8_t> bits_t;
 
 private:
-  int    m_num_bits;
+  ulong    m_num_bits;
   bits_t m_bits;
 
 public:
   DynamicBitset() { }
 
-  DynamicBitset(int size)
+  DynamicBitset(ulong size)
     :
     m_num_bits(size),
     m_bits((size >> 3) + ((size & 7) ? 1 : 0)) 
@@ -49,20 +49,20 @@ public:
 
   ~DynamicBitset() { }
 
-  int size() const
+  ulong size() const
   {
     return m_num_bits;
   }
 
-  void getSet(std::vector<int>& set) const
+  void getSet(std::vector<ulong>& set) const
   {
-    int numBytes = m_bits.size();
-    for (int i = 0; i < numBytes; ++i)
+    ulong numBytes = m_bits.size();
+    for (ulong i = 0; i < numBytes; ++i)
     {
       if (m_bits[i] == 0)
         continue;
 
-      for (int j = 0; j < 8; ++j) 
+      for (ulong j = 0; j < 8; ++j) 
       {
         if ((m_bits[i] & (128 >> j)) != 0) // 128 = 1000 0000
           set.push_back(i * 8 + j);
@@ -81,16 +81,16 @@ public:
     return (byte & (128 >> getBitWithinByte(bit))) != 0;
   }
 
-  void set(int bit, bool value)
+  void set(ulong bit, bool value)
   {
-    int off = getByteIndex(bit);
+    ulong off = getByteIndex(bit);
     if (!value && m_bits.size() <= (unsigned)off)
       return;
 
     if (m_bits.size() <= (unsigned)off)
       m_bits.resize(off + 1);
  
-    int bitnum = getBitWithinByte(bit);
+    ulong bitnum = getBitWithinByte(bit);
     if (value) 
     {
       m_bits[off] |= (128 >> bitnum);
@@ -112,17 +112,17 @@ public:
   
   DynamicBitset& set_union(const DynamicBitset& rhs)
   {
-    int idx = 0;
-    while((unsigned)idx < m_bits.size() && (unsigned)idx < rhs.m_bits.size()) 
+    ulong idx = 0;
+    while(idx < m_bits.size() && idx < rhs.m_bits.size()) 
     {
       m_bits[idx] |= rhs.m_bits[idx];
       ++idx;
     }
 
-    if ((unsigned)idx < rhs.m_bits.size()) 
+    if (idx < rhs.m_bits.size()) 
     {
       m_bits.resize(rhs.m_bits.size());
-      while((unsigned)idx < rhs.m_bits.size()) 
+      while(idx < rhs.m_bits.size()) 
       {
         m_bits.push_back(rhs.m_bits[idx++]);
       }
@@ -132,13 +132,13 @@ public:
   
   DynamicBitset& set_intersect(const DynamicBitset& rhs)
   {
-    int idx = 0;
-    while((unsigned)idx < m_bits.size() && (unsigned)idx < rhs.m_bits.size()) 
+    ulong idx = 0;
+    while(idx < m_bits.size() && idx < rhs.m_bits.size()) 
     {
       m_bits[idx] &= rhs.m_bits[idx];
       ++idx;
     }
-    while((unsigned)idx < m_bits.size()) 
+    while(idx < m_bits.size()) 
     {
       m_bits[idx++] = 0;
     }
@@ -147,12 +147,12 @@ public:
 
 
 private:
-  static int getByteIndex(int bit)
+  static ulong getByteIndex(ulong bit)
   {
     return bit >> 3;
   }
   
-  static int getBitWithinByte(int bit)
+  static ulong getBitWithinByte(ulong bit)
   {
     // bit = 8 ==> bit & 7 = 0,
     // bit = 9 ==> bit & 7 = 1, etc
